+++ plugin-dev/Source/Sentry/Private/GenericPlatform/GenericPlatformSentrySubsystem.h
@@ -98,6 +98,11 @@ private:
 	FCriticalSection CriticalSection;
 
 	FString databaseParentPath;
+
+	/**
+	 * The expected destination URL of the crash envelope.
+	 */
+	FString EventDestinationUrl;
 };
 
 #endif